Improve task cleanup tests
