+++ tests/utils/io.py
@@ -1,6 +1,11 @@
 import socket
 from contextlib import closing
 
+from . import wait_for_job_state
+from .job import default_task_output
+from .wait import wait_until
+from ..conftest import HqEnv
+
 
 def check_file_contents(path: str, content):
     assert read_file(path) == str(content)
@@ -25,3 +30,18 @@ def find_free_port():
         s.bind(("", 0))
         s.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
         return s.getsockname()[1]
+
+
+def read_task_pid(hq_env: HqEnv, job_id: int = 1) -> int:
+    """
+    Reads the task's PID from its stdout.
+    """
+    wait_for_job_state(hq_env, job_id, "RUNNING")
+
+    def get_pid():
+        pid = read_file(default_task_output(job_id=job_id)).strip()
+        if not pid:
+            return None
+        return int(pid)
+
+    return wait_until(get_pid)
